[gws][add] job log
