+++ app/jobs/concerns/job/ss/loggable.rb
@@ -8,7 +8,7 @@ module Job::SS::Loggable
   private
 
   def perform_with_job_log
-    job_log = Job::Log.create_from_active_job(self)
+    job_log = create_job_log
     Job::TaskLogger.attach(job_log)
     ret = nil
     begin
@@ -34,6 +34,10 @@ module Job::SS::Loggable
       job_log.save
     end
     ret
+  end
+
+  def create_job_log
+    Job::Log.create_from_active_job(self)
   end
 
   def system_error?(e)
