+++ app/jobs/gws/application_job.rb
@@ -1,5 +1,5 @@
 class Gws::ApplicationJob < ::ApplicationJob
   include Job::SS::Core
   include Job::Gws::Binding::Base
-  # include Job::Gws::Loggable
+  include Job::Gws::Loggable
 end
