+++ app/jobs/gws/reminder/notification_job.rb
@@ -6,11 +6,14 @@ class Gws::Reminder::NotificationJob < Gws::ApplicationJob
   def perform(opts = {})
     from = opts[:from] || now - 10.minutes
     to = opts[:to] || now
-    reminder_ids = Gws::Reminder.site(site).notify_around(from, to).pluck(:id)
+    reminder_ids = Gws::Reminder.site(site).notify_between(from, to).pluck(:id)
     reminder_ids.each do |reminder_id|
       item = Gws::Reminder.find(reminder_id)
       mail = Gws::Reminder::Mailer.notify_mail(item)
-      mail.deliver_now if mail.present?
+      if mail.present?
+        Rails.logger.info("#{mail.to.first}: send reminder mail")
+        mail.deliver_now
+      end
     end
   end
 end
