+++ spec/jobs/gws/reminder/notification_job_spec.rb
@@ -35,5 +35,11 @@ describe Gws::Reminder::NotificationJob, dbscope: :example do
       expect(notify_mail.body.raw_source).to include("[参加ユーザー]\n")
       expect(notify_mail.body.raw_source).to include(schedule.members.first.long_name)
     end
+
+    expect(Gws::Job::Log.count).to eq 1
+    Gws::Job::Log.first.tap do |log|
+      expect(log.logs).to include(include("INFO -- : Started Job"))
+      expect(log.logs).to include(include("INFO -- : Completed Job"))
+    end
   end
 end
