+++ spec/models/gws/schedule/plan_spec.rb
@@ -73,41 +73,4 @@ RSpec.describe Gws::Schedule::Plan, type: :model, dbscope: :example, tmpdir: tru
       end
     end
   end
-
-  describe "notification" do
-    let(:schedule) { create :gws_schedule_plan }
-    let(:reminder) { Gws::Reminder.where(item_id: schedule.id, model: described_class.name.underscore).first }
-
-    before do
-      notification = reminder.notifications.new(in_notify_before: 30)
-      notification.valid?
-      reminder.save!
-    end
-
-    before do
-      ActionMailer::Base.deliveries = []
-    end
-
-    after do
-      ActionMailer::Base.deliveries = []
-    end
-
-    it do
-      Timecop.travel(reminder.notifications.first.notify_at) do
-        Gws::Reminder.send_notification_mail(Time.zone.now - 9.minutes, Time.zone.now + 1.minute)
-      end
-
-      expect(ActionMailer::Base.deliveries.length).to eq 1
-      ActionMailer::Base.deliveries.first.tap do |notify_mail|
-        expect(notify_mail.from.first).to eq reminder.user.email
-        expect(notify_mail.to.first).to eq reminder.user.email
-        expect(notify_mail.subject).to eq "[リマインダー] スケジュール - #{schedule.name}"
-        expect(notify_mail.body.multipart?).to be_falsey
-        expect(notify_mail.body.raw_source).to include("[タイトル] #{schedule.name}")
-        expect(notify_mail.body.raw_source).to include("[日時] #{I18n.l(schedule.start_at.to_date, format: :gws_long)}")
-        expect(notify_mail.body.raw_source).to include("[参加ユーザー]\n")
-        expect(notify_mail.body.raw_source).to include(schedule.members.first.long_name)
-      end
-    end
-  end
 end
